Cleaned up hashFiles function
Originally the for loop in the hashFiles function used "i" as the
variable that iterated over the list of files. Immediately after
the for loop was created the "name" variable was created and set to
"i". This was redundent so "i" was changed to "name" and every instance
of "i" was replaced with "name"