added test folder with files
a test folder called "test_files" was added in the dupfileremover
directory. It has two directories in it. They contain files that
are both copies and unique in each folder. If a file ends with
"_cp" then the original is in the opposite folder. If a file ends
with "_uniq" then that is a unique file that exists only in that directory.
These files can be used for testing purposes to ensure the program is
behaving as expected