Created command line functionality using click
Before the program was a simple python script in a single python file.
Parsing the arguments that were passed to it was extreamly clunky, not
graceful at all and was easily broken. It has now been updated using
click to parse all arguments and options given to the program (see
below for link to click documentation). click uses decorators to
manage both required arguments and options that the program will use.
A new file (command_line.py) was created to manage the flow of the
program using click. duplicateChecker, which was originally the
script itself, is now just used as a module while click manages all
inputs from the user. setup.py was edited so the command line
entry point is now the "cli" function within "command_line.py"

click documentation: http://click.pocoo.org/5/