Added Progress Bar
A progress bar was added to show how far along the process of hashing
the files is. It was implemented by tqdm (link below). All that
needed to be done in order to do it was to wrap the "fileList"
variable in the "hashFiles" function under "duplicateChecker.py" with
tqdm().

tqdm: https://pypi.python.org/pypi/tqdm